// electro-pime/types/api.ts
export interface Customer {
  id: string;
  name: string;
  email?: string;
  phone: string;
  documentType: string;
  documentNumber: string;
  address?: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCustomerDto {
  name: string;
  email?: string;
  phone: string;
  documentType: string;
  documentNumber: string;
  address?: string;
}

// ==================== Product ====================

export interface Product {
  id: string;
  name: string;
  description?: string;
  price: number;
  cost: number;
  stock: number;
  category: string;
  isActive: boolean;
  imageUrl?: string;
  createdAt: string;
  updatedAt: string;
}
// ...
export type PaymentMethod = 'CASH' | 'CREDIT_CARD' | 'DEBIT_CARD' | 'TRANSFER' | 'YAPE' | 'PLIN';
// ...
export interface CreateSaleDto {
  customerId?: string;
  customerName?: string;
  paymentMethod: PaymentMethod;
  items: {
    productId: string;
    quantity: number;
    price: number;
  }[];
}